TA-3316 updated resources that use cloudsearch to use single implementation
