+++ src/service/office-search.js
@@ -1,58 +1,12 @@
 const config = require('../config')
-const aws = require('aws-sdk')
-const csd = new aws.CloudSearchDomain({
-  endpoint: config.cloudSearch.officeEndpoint,
-  region: 'us-east-1',
-  apiVersions: '2013-01-01'
-})
+const cloudsearch = require('../clients/cloudsearch.js')
+const location = require('./location.js')
 
+const endpoint = config.cloudSearch.officeEndpoint
 const defaultOfficeType = 'SBA district office'
 const defaultOfficeGeocode = {
   latitude: 38.893311,
   longitude: -77.014647
-}
-let location = require('./location.js')
-
-// for testing purposes
-async function runSearch (params) {
-  const result = await csd.search(params).promise()
-  return result
-}
-
-// function buildFilterQuery(req) {
-//   const filters = buildFilters(req)
-//   const filterArray = []
-//   //todo use lodash to filter out empty filters
-//   for (const filter of filters) {
-//     if (filter) {
-//       filterArray.push(filter)
-//     }
-//   }
-//   return filterArray.length ? `(and ${filterArray.join(' ')})` : null
-// }
-
-// function buildIsSbaOfficeFilterQuery(isSbaOffice) {
-//   let sbaOfficeString = ''
-//   if (isSbaOffice && (isSbaOffice === true || isSbaOffice.toLowerCase() === 'true')) {
-//     const sbaOfficeNames = config.get('features.office.sbaOfficeNames')
-//     const sbaSearchStrings = sbaOfficeNames.map(officeName => {
-//       return `office_type: '${formatString(officeName)}'`
-//     })
-//     if (sbaOfficeNames.length === 1) {
-//       sbaOfficeString = sbaSearchStrings[0]
-//     } else if (sbaOfficeNames.length > 1) {
-//       sbaOfficeString = `(or ${sbaSearchStrings.join(' ')})`
-//     }
-//   }
-//   return sbaOfficeString
-// }
-
-function formatString (string) {
-  let result = decodeURI(string)
-  // cloudsearch requires us to escape backslashes and quotes
-  result = result.replace(/\\/g, '\\\\')
-  result = result.replace(/'/g, "\\'")
-  return result
 }
 
 function buildQuery (query) {
@@ -62,7 +16,7 @@ function buildQuery (query) {
   if (query) {
     const fieldsToSearch = ['title', 'location_name', 'office_type']
     for (const field of fieldsToSearch) {
-      queryStatements.push(`${field}: '${formatString(query)}'`)
+      queryStatements.push(`${field}: '${cloudsearch.formatString(query)}'`)
     }
   }
   if (queryStatements.length > 1) {
@@ -73,8 +27,8 @@ function buildQuery (query) {
 
 function buildFilters (service, type) {
   let filters = [
-    service ? `office_service: '${formatString(service)}'` : null,
-    type ? `office_type: '${formatString(type)}'` : null
+    service ? `office_service: '${cloudsearch.formatString(service)}'` : null,
+    type ? `office_type: '${cloudsearch.formatString(type)}'` : null
   ]
   filters = filters.filter(item => item)
   let filterString = null
@@ -141,7 +95,7 @@ async function fetchOffices (query) {
   let geo = await location.generateGeocode(address, mapCenter)
   const params = buildParams(queryObj, geo)
   try {
-    const result = await module.exports.runSearch(params) // call the module.exports version for stubbing during testing
+    const result = await cloudsearch.runSearch(params, endpoint) // call the module.exports version for stubbing during testing
     const hits = result.hits
     if (hits && hits.found > 0) {
       const newHitList = hits.hit.map(item => {
@@ -159,7 +113,7 @@ async function fetchOffices (query) {
       return Object.assign({}, hits, { hit: newHitList })
     } else {
       const defaultParams = buildDefaultOfficeQueryParams(geo)
-      const suggestedResults = await module.exports.runSearch(defaultParams)
+      const suggestedResults = await cloudsearch.runSearch(defaultParams, endpoint)
       return Object.assign({}, hits, { suggestedResults: suggestedResults.hits })
     }
   } catch (err) {
@@ -169,6 +123,3 @@ async function fetchOffices (query) {
 }
 
 module.exports.fetchOffices = fetchOffices
-
-// for testing purposes
-module.exports.runSearch = runSearch
